refactor(login): move login components in a directory
